+++ src/main/java/exceptions/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package exceptions;
 
 public class InvalidCommandException extends Exception {
