+++ src/main/java/exceptions/InvalidIDException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package exceptions;
 
 public class InvalidIDException extends Exception {
     public InvalidIDException(String message) {
