+++ src/main/java/tasks/Events.java
@@ -1,4 +1,4 @@
-package TaskPackage;
+package tasks;
 
 public class Events extends TasksDefault {
