+++ src/main/java/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package TaskPackage;
+package tasks;
 
 public class ToDo extends TasksDefault {
